move coordinates calculation to separate module
